+++ src/main/java/lotto/Application.java
@@ -1,7 +1,10 @@
 package lotto;
 
+import lotto.controller.LottoVender;
+
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
+        LottoVender lottovender = new LottoVender();
     }
 }
