+++ src/test/java/lottovender/LottoVenderTest.java
@@ -1,53 +1,45 @@
 package lottovender;
 
-import lotto.Lotto;
+import lotto.controller.LottoVender;
+import lotto.domain.Lotto;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+
+import java.util.ArrayList;
 import java.util.List;
 
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
 
 class LottoVenderTest {
 
-    @DisplayName("구매 금액을 빈칸으로 입력할 시 예외 발생")
-    @Test
-    void inputBlank() {
-        LottoVender lv = new LottoVender();
-        Assertions.assertThrows(IllegalArgumentException.class, () -> {
-            lv.payPrice(" ");
-        });
-    }
-
-    @DisplayName("숫자가 아닐경우 예외 처리")
-    @Test
-    void input_not_a_number() {
-        LottoVender lv = new LottoVender();
-        Assertions.assertThrows(IllegalArgumentException.class, () -> {
-            lv.payPrice("error");
-        });
-    }
-
-    @DisplayName("숫자가 아닐경우 예외 처리")
-    @Test
-    void input_can_not_be_divided_by_1000() {
-        LottoVender lv = new LottoVender();
-        Assertions.assertThrows(IllegalArgumentException.class, () -> {
-            lv.payPrice("1004");
-        });
-    }
-
-    @DisplayName("로또 구입 비용 지불")
+    @DisplayName("지불한 금액 만큼 로또 발행수 카운트")
     @Test
     void pay_price_for_lotto() {
         LottoVender lv = new LottoVender();
-        Assertions.assertTrue(lv.payPrice("15000")==15);
+        lv.payPrice(15000);
+        Assertions.assertTrue(lv.getLottoCount()==15);
     }
 
+    @DisplayName("로또 발행")
     @Test
-    void getLotto() {
-        System.out.println("1");
+    void get_lotto() {
+        //비교 대상 로또
+        List<Integer> testNumbers = new ArrayList<>();
+        testNumbers.add(3);
+        testNumbers.add(5);
+        testNumbers.add(14);
+        testNumbers.add(17);
+        testNumbers.add(22);
+        testNumbers.add(41);
+        Lotto testLotto = new Lotto(testNumbers);
+
+        LottoVender lv = new LottoVender();
+        List<Lotto> lottoticekts = lv.getLotto(1);
+
+        for(Lotto lotto:lottoticekts){
+            assertThat(lotto).isEqualTo(testLotto);
+        }
     }
 }
